install.sh: Fetch latest version & add linux aarch64 support (#11)
* Implement `get_latest_release` in `install.sh`

This commit updates the `install.sh` sript from wasmtime to download the
latest stable release instead of the `dev` release. This should be a bit
more stable over time since the `dev` release is rebuilt on all CI runs
for the Wasmtime repository and the versioned releases should be pretty
stable nowadays as well.

* Specifically support aarch64 Linux in install.sh

We don't have full platform support for aarch64 on all platforms but we
do have specific support for aarch64 Linux.