Add support for aarch64 and s390x binaries.
This fixes bytecodealliance/wasmtime#4009. It adds support for
downloading and installing aarch64 and s390x binaries on Linux.

(Note that while #11 added Linux/aarch64 to the allowlist, the download
URL still had `x86_64` hardcoded; this PR fixes that.)

It also adds most of what is needed for macOS/aarch64, once we release
v0.37.0; I verified that, with the one-line change to allowlist the
platform, the script hits the correct URL on my M1 macOS system (and
gets a 404 since v0.36.0 does not include M1 binaries). I'll do a
separate PR to update this after the release.

Tested on Linux/aarch64, Linux/x86_64, macOS/aarch64 (as noted above,
ready for next release) and macOS/x86_64.