Add RISC-V support to install.sh
This commit adds support for the RISC-V Architecture
to the install.sh script.

Wasmtime already publishes precompiled binaries for
this target, we just need to add some minimal support
to get it working here.

There is a slight wrinkle, Linux identifies RISC-V
as "riscv64", but wasmtime uses the name "riscv64gc".
So this commit adds a mapping from the former to
the latter when fetching the binaries.